# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Subscrition Platform APi",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/posts": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "List all posts",
                "operationId": "posts-listing",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "422": {
                        "description": "Validation Failed"
                    }
                }
            },
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "Create a new post",
                "operationId": "post-create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "website_id",
                                    "title"
                                ],
                                "properties": {
                                    "website_id": {
                                        "description": "Associate Website",
                                        "type": "text"
                                    },
                                    "title": {
                                        "description": "Post Title",
                                        "type": "text"
                                    },
                                    "description": {
                                        "description": "Post description",
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "422": {
                        "description": "Validation Failed"
                    }
                }
            }
        },
        "/api/v1/posts/{id}": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "view a post",
                "operationId": "post-view",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "422": {
                        "description": "Validation Failed"
                    }
                }
            }
        },
        "/api/v1/subscriptions": {
            "post": {
                "tags": [
                    "subscriptions"
                ],
                "summary": "Create a new subscriber",
                "operationId": "subscriber-create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "website_id",
                                    "user_id"
                                ],
                                "properties": {
                                    "website_id": {
                                        "description": "Associate Website",
                                        "type": "text"
                                    },
                                    "user_id": {
                                        "description": "Associate User",
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "422": {
                        "description": "Validation Failed"
                    }
                }
            }
        },
        "/api/v1/websites": {
            "get": {
                "tags": [
                    "WebSites"
                ],
                "summary": "List all websites",
                "operationId": "websites-listing",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "422": {
                        "description": "Validation Failed"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "jwt": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}